requirements.txt
alabaster==0.7.10
anglerfish==1.7.5
appdirs==1.4.3
arrow==0.10.0
astroid==1.4.9
Babel==2.3.4
beautifulsoup4==4.5.3
binaryornot==0.4.0
bs4==0.0.1
chardet==2.3.0
click==6.7
colorama==0.3.7
cookiecutter==1.5.1
coverage==4.1
css-html-js-minify==2.2.2
cycler==0.10.0
decorator==4.0.11
defusedxml==0.4.1
Django==1.10.6
django-allauth==0.24.1
django-bootstrap3==6.2.2
django-countries==3.4.1
django-extensions==1.6.1
django-postman==3.3.1
djangorestframework==3.3.2
docutils==0.13.1
Flask==0.11.1
future==0.16.0
graphviz==0.6
imagesize==0.7.1
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.5
jinja2-time==0.2.0
lazy-object-proxy==1.2.2
MarkupSafe==1.0
matplotlib==2.0.0
mccabe==0.6.1
numpy==1.12.1
oauthlib==1.0.3
olefile==0.44
packaging==16.8
pickleshare==0.7.4
Pillow==4.0.0
poyo==0.4.1
prompt-toolkit==1.0.13
psycopg2==2.6.1
py==1.4.33
py2exe==0.9.2.2
PyExecJS==1.3.0
pyglet==1.2.4
Pygments==2.2.0
pylint==1.6.5
pyparsing==2.2.0
pypiwin32==219
PyReact==0.6.0
PyScaffold==2.5.7
pyserial==3.2.1
pytest==3.0.7
pytest-cov==2.4.0
python-dateutil==2.6.0
python3-openid==3.0.9
pytz==2016.10
pywinusb==0.4.1
requests==2.13.0
requests-oauthlib==0.6.0
scikit-learn==0.18
scipy==0.18.1
simplegeneric==0.8.1
sip==4.18.1
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.5.3
traitlets==4.3.2
Twisted==16.3.0
wcwidth==0.1.7
Werkzeug==0.11.11
whichcraft==0.4.0
win-unicode-console==0.5
wrapt==1.10.10
zope.interface==4.2.0
